Finally got bullet sprites loading properly.
Moved the BulletTex and PlayerTex out into the Main class, so that they exist for as long as the game does---then just pass them to the player and bullet constructors as pointers to them---de reference them and assign to sprite of bullet/player/whatever.

Maybe I could make a class to handle the textures? I dno.