Player Movement tuned.
Removed cases where diagonal inputs would cause maintanance of one of the axis of velocity even when one of the diagonal input movement keys was released. IE the case of pressing W+A and then releasing A but maintaining the W axis of movement due to the air resistance only kicking in once keys pressed = 0.